Refactor raw message test into gobotintegrationtests
